Separate React Router 6 and 6.4 tests.
